+++ app/src/main/java/eu/kanade/domain/entries/manga/interactor/UpdateManga.kt
@@ -72,6 +72,7 @@ class UpdateManga(
                 status = remoteManga.status.toLong(),
                 updateStrategy = remoteManga.update_strategy,
                 initialized = true,
+                memo = remoteManga.memo,
             ),
         )
     }