+++ domain/src/main/java/tachiyomi/domain/items/chapter/model/Chapter.kt
@@ -1,5 +1,8 @@
 package tachiyomi.domain.items.chapter.model
 
+import eu.kanade.tachiyomi.source.model.EMPTY
+import kotlinx.serialization.json.JsonObject
+
 data class Chapter(
     val id: Long,
     val mangaId: Long,
@@ -15,6 +18,7 @@ data class Chapter(
     val scanlator: String?,
     val lastModifiedAt: Long,
     val version: Long,
+    val memo: JsonObject = JsonObject.EMPTY,
 ) {
     val isRecognizedNumber: Boolean
         get() = chapterNumber >= 0f
@@ -26,6 +30,7 @@ data class Chapter(
             dateUpload = other.dateUpload,
             chapterNumber = other.chapterNumber,
             scanlator = other.scanlator?.ifBlank { null },
+            memo = other.memo,
         )
     }